+++ Backend/controllers/post.js
@@ -64,19 +64,32 @@ exports.updateOne = (req, res, next) => {
 };
 
 exports.deletePublish = (req, res, next) => {
-    let post_id = req.params.idPost;;
+    let post_id = req.params.idPost;
+    let id_user = req.params.id;
+    
 
-    let sqlDelete = `DELETE FROM post WHERE id='${post_id}'`;
-    db.query(sqlDelete, (err, result) => {
-        if (err) {
+    let sqlCheck = `SELECT user_id FROM post WHERE id='${post_id}'`;
+    db.query(sqlCheck, (err, result) => {
+        if (result[0].user_id != id_user) {
             return res.status(404).json({
-                message: "Supression erreur"
+                message: "Supression non authorisé"
             });
-        };
-        res.status(200).json({
-            message: "Post suprimé"
-        });
+        }else{
+            let sqlDelete = `DELETE FROM post WHERE id='${post_id}'`;
+            db.query(sqlDelete, (err, result) => {
+                if (err) {
+                    return res.status(404).json({
+                        message: "Supression erreur"
+                    });
+                };
+                res.status(200).json({
+                    message: "Post suprimé"
+                });
+            })
+        }
     })
+
+    
 };
 
 exports.allPublish = (req, res, next) => {
